Introduce Panorama support and refactor device management
Add Panorama client initialization and device retrieval functions to `devices/panorama.go`. Remove Panorama-specific code from `devices/devices.go` and relocate NGFW-related functions to `devices/ngfw.go`. This modularizes the device management logic, separating concerns for better clarity and maintainability.
